requirements.txt
absl-py==0.9.0
astor==0.8.1
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
Cython==0.29.15
future==0.18.2
gast==0.2.2
get==2019.4.13
google-auth==1.11.2
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.27.2
h5py==2.10.0
idna==2.9
imageai==2.1.5
imutils==0.5.3
iso8601==0.1.12
Keras==2.1.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.2.1
matplotlib==3.2.0
mysql-connector==2.2.9
numpy==1.14.4
oauthlib==3.1.0
opencv-python==4.1.0.25
opt-einsum==3.2.0
Pillow==7.0.0
post==2019.4.13
protobuf==3.11.3
public==2019.4.13
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.6
pyserial==3.4
pytesseract==0.3.2
python-dateutil==2.8.1
PyYAML==5.3
query-string==2019.4.13
request==2019.4.13
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
six==1.14.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
tesseract==0.1.3
urllib3==1.25.8
Werkzeug==1.0.0
wrapt==1.12.0
yolov3==1.7
